add Message object, remove validation error handler, and imporve validation error collection
...

fix tests

refactor GenericFailure

...

...

add post_init check to GenericFailure

...

...

...
