+++ app.py
@@ -12,8 +12,7 @@ from werkzeug.serving import WSGIRequestHandler
 from config import Config
 from core.cli import create_cli
 from core.db import db, migrate
-from core.exceptions import ValidationError
-from core.handlers import handle_exception, handle_validation_error
+from core.handlers import handle_exception
 
 WORKING_DIR = Path(__file__).parent
 
@@ -47,7 +46,6 @@ def create_app(config_class=Config) -> Flask:
     )
     flask_app.logger.info(f"Database: {str(flask_app.config.get('SQLALCHEMY_DATABASE_URI')).split('://')[0]}")
 
-    flask_app.register_error_handler(ValidationError, handle_validation_error)
     flask_app.register_error_handler(Exception, handle_exception)
 
     create_cli(flask_app)
