+++ core/const.py
@@ -707,182 +707,6 @@ TABLE_SORT_ORDERS = {
     "RiskRegister": ["SubmissionID", "ProgrammeID", "ProjectID", "RiskName"],
 }
 
-# Internal table names to Round 3 & 4 TF tab names mapping
-INTERNAL_TABLE_TO_FORM_SHEET = {
-    "Project Details": "Project Admin",
-    "Project Progress": "Programme Progress",
-    "Programme Progress": "Programme Progress",
-    "Funding": "Funding Profiles",
-    "Funding Questions": "Funding Profiles",
-    "Outcome_Data": "Outcomes",
-    "Output_Data": "Project Outputs",
-    "RiskRegister": "Risk Register",
-    "Place Details": "Project Admin",
-    "Private Investments": "PSI",
-}
-
-# Internal column names to Round 3 & 4 TF column and section mapping
-INTERNAL_COLUMN_TO_FORM_COLUMN_AND_SECTION = {
-    "Single or Multiple Locations": (
-        "Does the project have a single location (e.g. one site) or multiple (e.g. multiple sites or across a number "
-        "of post codes)?",
-        "Project Details",
-    ),
-    "GIS Provided": ("Are you providing a GIS map (see guidance) with your return?", "Project Details"),
-    "Answer": ("Answer", "Programme-Wide Progress Summary"),
-    "Start Date": ("Start Date - mmm/yy (e.g. Dec-22)", "Projects Progress Summary"),
-    "Completion Date": ("Completion Date - mmm/yy (e.g. Dec-22)", "Projects Progress Summary"),
-    "Current Project Delivery Stage": ("Current Project Delivery Stage", "Projects Progress Summary"),
-    "Project Adjustment Request Status": ("Project Adjustment Request Status", "Projects Progress Summary"),
-    "Project Delivery Status": ("Project Delivery Status", "Projects Progress Summary"),
-    "Leading Factor of Delay": ("Leading Factor of Delay", "Projects Progress Summary"),
-    "Delivery (RAG)": ("Delivery (RAG)", "Projects Progress Summary"),
-    "Spend (RAG)": ("Spend (RAG)", "Projects Progress Summary"),
-    "Risk (RAG)": ("Risk (RAG)", "Projects Progress Summary"),
-    "Commentary on Status and RAG Ratings": ("Commentary on Status and RAG Ratings", "Projects Progress Summary"),
-    "Most Important Upcoming Comms Milestone": ("Most Important Upcoming Comms Milestone", "Projects Progress Summary"),
-    "Date of Most Important Upcoming Comms Milestone (e.g. Dec-22)": (
-        "Date of Most Important Upcoming Comms Milestone (e.g. Dec-22)",
-        "Projects Progress Summary",
-    ),
-    "Secured": ("Has this funding source been secured?", "Project Funding Profiles"),
-    "GeographyIndicator": ("Geography Indicator", "Outcome Indicators (excluding footfall)"),
-    "RiskName": ("Risk Name", "Programme / Project Risks"),
-    "RiskCategory": ("Risk Category", "Programme / Project Risks"),
-    "Short Description": ("Short description of the Risk", "Programme / Project Risks"),
-    "Full Description": ("Full Description", "Programme / Project Risks"),
-    "Consequences": ("Consequences", "Programme / Project Risks"),
-    "Pre-mitigatedImpact": ("Pre-mitigated Impact", "Programme / Project Risks"),
-    "Pre-mitigatedLikelihood": ("Pre-mitigated Likelihood", "Programme / Project Risks"),
-    "Mitigatons": ("Mitigations", "Programme / Project Risks"),
-    "PostMitigatedImpact": ("Post-Mitigated Impact", "Programme / Project Risks"),
-    "PostMitigatedLikelihood": ("Post-mitigated Likelihood", "Programme / Project Risks"),
-    "Proximity": ("Proximity", "Programme / Project Risks"),
-    "RiskOwnerRole": ("Risk Owner/Role", "Programme / Project Risks"),
-    "Funding Source Name": ("Funding Source Name", "Project Funding Profiles"),
-    "Funding Source Type": ("Funding Source", "Project Funding Profiles"),
-    "Start_Date": ("H1 (Apr-Sep)", "Project Funding Profiles"),
-    "End_Date": ("H2 (Oct-Mar)", "Project Funding Profiles"),
-    "Total Project Value": ("Total Project Value (£)", "Private Sector Investment"),
-    "Townsfund Funding": ("Award From Townsfund (£)", "Private Sector Investment"),
-    "Output": ("Indicator Name", "Project Outputs"),
-    "Unit of Measurement": ("Unit of Measurement", "Project Outputs"),
-    "UnitofMeasurement": ("Unit of Measurement", "Outcome Indicators (excluding footfall) / Footfall Indicator"),
-    "Outcome": ("Indicator Name", "Outcome Indicators (excluding footfall) / Footfall Indicator"),
-    "Project Name": ("Project Name", "Project Details"),
-    "Primary Intervention Theme": ("Primary Intervention Theme", "Project Details"),
-    "Locations": ("Project Location(s) - Post Code (e.g. SW1P 4DF)", "Project Details"),
-    "Lat/Long": ("Project Location - Lat/Long Coordinates (3.d.p e.g. 51.496, -0.129)", "Project Details"),
-    "Private Sector Funding Required": ("Private Sector Funding Required", "Private Sector Investment"),
-    "Private Sector Funding Secured": ("Private Sector Funding Secured", "Private Sector Investment"),
-    "Spend for Reporting Period": ("Financial Year 2022/21 - Financial Year 2025/26", "Project Funding Profiles"),
-    "Amount": ("Financial Year 2022/21 - Financial Year 2025/26", "Project Outputs"),
-}
-
-# form version and reporting period for different rounds
-GET_FORM_VERSION_AND_REPORTING_PERIOD = {
-    3: ("Town Deals and Future High Streets Fund Reporting Template (v3.0)", "1 October 2022 to 31 March 2023"),
-    4: (
-        f"Town Deals and Future High Streets Fund Reporting Template ({TF_ROUND_4_TEMPLATE_VERSION})",
-        "1 April 2023 to 30 September 2023",
-    ),
-}
-
-INTERNAL_TYPE_TO_MESSAGE_FORMAT = {
-    "datetime64[ns]": "a date",
-    "float64": "a number",
-    "string": "text",
-    "int64": "a number",
-    "object": "an unknown datatype",
-}
-
-PRE_DEFINED_FUNDING_SOURCES = [
-    "Commercial Income",
-    "How much of your CDEL forecast is contractually committed?",
-    "How much of your RDEL forecast is contractually committed?",
-    "Town Deals 5% CDEL Pre-Payment",
-    "Towns Fund CDEL which is being utilised on TF project related activity (For Town Deals, this excludes the 5% CDEL "
-    "Pre-Payment)",
-    "Towns Fund RDEL Payment which is being utilised on TF project related activity",
-]
-
-# mapping of user submitted column names per table to its original excel column letter index
-# for the Towns Fund round 4 spreadsheet
-TABLE_AND_COLUMN_TO_ORIGINAL_COLUMN_LETTER = {
-    "Place Details": {"Question": "C{i}", "Indicator": "D{i}", "Answer": "E{i}"},
-    "Project Details": {
-        "Project name": "E{i}",
-        "Primary Intervention Theme": "F{i}",
-        "Single or Multiple Locations": "G{i}",
-        "Locations": "H{i} or K{i}",
-        "Postcodes": "H{i} or K{i}",
-        "Lat/Long": "I{i} or L{i}",
-        "GIS Provided": "J{i}",
-    },
-    "Programme Progress": {"Question": "C{i}", "Answer": "D{i}"},
-    "Project Progress": {
-        "Start Date": "D{i}",
-        "Completion Date": "E{i}",
-        "Current Project Delivery Stage": "F{i}",
-        "Project Delivery Status": "G{i}",
-        "Leading Factor of Delay": "H{i}",
-        "Project Adjustment Request Status": "I{i}",
-        "Delivery (RAG)": "J{i}",
-        "Spend (RAG)": "K{i}",
-        "Risk (RAG)": "L{i}",
-        "Commentary on Status and RAG Ratings": "M{i}",
-        "Most Important Upcoming Comms Milestone": "N{i}",
-        "Date of Most Important Upcoming Comms Milestone (e.g. Dec-22)": "O{i}",
-    },
-    "Funding Questions": {
-        "All Columns": "E{i}",  # stretches across all 3 columns below
-        "TD 5% CDEL Pre-Payment\n(Towns Fund FAQs p.46 - 49)": "E{i}",
-        "TD RDEL Capacity Funding": "F{i}",
-        "TD Accelerated Funding": "I{i}",
-    },
-    "Funding Comments": {"Comment": "C{i} to E{i}"},
-    "Funding": {
-        "Funding Source Name": "C{i}",
-        "Funding Source Type": "D{i}",
-        "Secured": "E{i}",
-        "Spend for Reporting Period": "F{i} to Y{i}",
-        "Grand Total": "Z{i}",
-    },
-    "Private Investments": {
-        "Private Sector Funding Required": "G{i}",
-        "Private Sector Funding Secured": "H{i}",
-        "Additional Comments": "J{i}",
-    },
-    "Output_Data": {
-        "Output": "C{i}",
-        "Unit of Measurement": "D{i}",
-        "Amount": "E{i} to W{i}",
-        "Additional Information": "Y{i}",
-    },
-    "Outcome_Data": {
-        "Outcome": "B{i}",
-        "UnitofMeasurement": "C{i}",
-        "Relevant project(s)": "D{i}",
-        "GeographyIndicator": "E{i}",
-        "Amount": "F{i} to O{i}",
-        "Higher Frequency": "P{i}",
-    },
-    "RiskRegister": {
-        "RiskName": "C{i}",
-        "RiskCategory": "D{i}",
-        "Short Description": "E{i}",
-        "Full Description": "F{i}",
-        "Consequences": "G{i}",
-        "Pre-mitigatedImpact": "H{i}",
-        "Pre-mitigatedLikelihood": "I{i}",
-        "Mitigatons": "K{i}",
-        "PostMitigatedImpact": "L{i}",
-        "PostMitigatedLikelihood": "M{i}",
-        "Proximity": "O{i}",
-        "RiskOwnerRole": "P{i}",
-    },
-}
-
 # map place to fund type associated with it
 PLACE_TO_FUND_TYPE = {
     "Heanor": {"Future_High_Street_Fund"},
@@ -1048,38 +872,4 @@ PLACE_TO_FUND_TYPE = {
     "Corby": {"Town_Deal"},
     "Northampton": {"Town_Deal", "Future_High_Street_Fund"},
     "Barrow": {"Town_Deal"},
-}
-
-
-FUND_TYPE_TO_TD_OR_HS = {"Future_High_Street_Fund": "HS", "Town_Deal": "TD"}
-
-
-# maps the financial year's start year back to original col letter for non-footfall outcomes
-FINANCIAL_YEAR_TO_ORIGINAL_COLUMN_LETTER_FOR_NON_FOOTFALL_OUTCOMES = {
-    2020: "F{i}",
-    2021: "G{i}",
-    2022: "H{i}",
-    2023: "I{i}",
-    2024: "J{i}",
-    2025: "K{i}",
-    2026: "L{i}",
-    2027: "M{i}",
-    2028: "N{i}",
-    2029: "O{i}",
-}
-
-# maps the month of the financial year's start year back to the original column for footfall outcomes
-MONTH_TO_ORIGINAL_COLUMN_LETTER_FOR_FOOTFALL_OUTCOMES = {
-    4: "D{i}",
-    5: "E{i}",
-    6: "F{i}",
-    7: "G{i}",
-    8: "H{i}",
-    9: "I{i}",
-    10: "J{i}",
-    11: "K{i}",
-    12: "L{i}",
-    1: "M{i}",
-    2: "N{i}",
-    3: "O{i}",
 }
