+++ core/validation/utils.py
@@ -4,10 +4,6 @@ from typing import Any
 
 import pandas as pd
 
-from core.const import (
-    FINANCIAL_YEAR_TO_ORIGINAL_COLUMN_LETTER_FOR_NON_FOOTFALL_OUTCOMES,
-    MONTH_TO_ORIGINAL_COLUMN_LETTER_FOR_FOOTFALL_OUTCOMES,
-)
 from core.validation.const import NA_VALUES
 
 
@@ -69,38 +65,6 @@ def is_from_dropdown(value: Any, dropdown_enum: StrEnum) -> bool:
     :return: True if blank, else False
     """
     return value in {dropdown_value for dropdown_value in dropdown_enum}
-
-
-def get_cell_indexes_for_outcomes(failed_row: pd.Series) -> str:
-    """
-    Constructs cell indexes for outcomes based on the provided failed row.
-
-    The function determines whether the error occurred in footfall outcomes (starting from row 60)
-    or non-footfall outcomes. If the index is greater than or equal to 60, it calculates the cell
-    index for footfall outcomes based on the provided 'Start_Date'. If the index is less than 60,
-    it calculates the cell index for non-footfall outcomes.
-
-    For footfall outcomes, the cell index is determined by adding a row index gap calculated from
-    the start year of the UK financial year as each year represents an additional 5 rows in the
-    original spreadsheet from the row where the entry for a given footfall outcome's data begins.
-
-    :param failed_row: A pandas Series representing a row where an error has occurred.
-    :return: A string containing the constructed cell index for outcomes.
-    """
-    start_date = failed_row["Start_Date"]
-    financial_year = get_uk_financial_year_start(start_date)
-    index = failed_row.name
-
-    # footfall outcomes starts from row 60
-    if index >= 60:
-        # row for 'Amount' column is end number of start year of financial year * 5 + 'Footfall Indicator' index
-        row_index_gap = int(str(financial_year)[-1]) * 5
-        index = int(index) + row_index_gap
-        cell_index = MONTH_TO_ORIGINAL_COLUMN_LETTER_FOR_FOOTFALL_OUTCOMES[start_date.month].format(i=index)
-    else:
-        cell_index = FINANCIAL_YEAR_TO_ORIGINAL_COLUMN_LETTER_FOR_NON_FOOTFALL_OUTCOMES[financial_year].format(i=index)
-
-    return cell_index
 
 
 def get_uk_financial_year_start(start_date: datetime) -> int:
