+++ tests/extraction_tests/test_round_four_extraction.py
@@ -188,18 +188,17 @@ def test_extract_outcomes_with_null_project(mock_outcomes_sheet, mock_project_lo
     mock_outcomes_sheet = mock_outcomes_sheet.replace("Test Project 1", np.nan)
     with pytest.raises(ValidationError) as ve:
         tf.extract_outcomes(mock_outcomes_sheet, mock_project_lookup, mock_programme_lookup, 4)
-    assert str(ve.value) == (
+    assert str(ve.value.validation_failures) == (
         (
-            "[GenericFailure(sheet='Outcomes', section='Outcome Indicators (excluding "
-            "footfall)', cell_index='D23', message='You’ve entered your own content, "
-            "instead of selecting from the dropdown list provided. Select an option from "
-            "the dropdown list.'), GenericFailure(sheet='Outcomes', section='Outcome "
-            "Indicators (excluding footfall)', cell_index='D24', message='You’ve entered "
-            "your own content, instead of selecting from the dropdown list provided. "
-            "Select an option from the dropdown list.'), GenericFailure(sheet='Outcomes', "
-            "section='Outcome Indicators (excluding footfall)', cell_index='D43', "
-            "message='You’ve entered your own content, instead of selecting from the "
-            "dropdown list provided. Select an option from the dropdown list.')]"
+            "[GenericFailure(table='Outcome_Data', section='Outcome Indicators (excluding footfall)', message='You’ve "
+            "entered your own content, instead of selecting from the dropdown list provided. Select an option from "
+            "the dropdown list.', cell_index=None, column='Relevant project(s)', row_index=23), GenericFailure("
+            "table='Outcome_Data', section='Outcome Indicators (excluding footfall)', message='You’ve entered your "
+            "own content, instead of selecting from the dropdown list provided. Select an option from the dropdown "
+            "list.', cell_index=None, column='Relevant project(s)', row_index=24), GenericFailure("
+            "table='Outcome_Data', section='Outcome Indicators (excluding footfall)', message='You’ve entered your "
+            "own content, instead of selecting from the dropdown list provided. Select an option from the dropdown "
+            "list.', cell_index=None, column='Relevant project(s)', row_index=43)]"
         )
     )
