+++ tests/test_handlers.py
@@ -4,51 +4,7 @@ from io import BytesIO
 
 from flask import g
 
-from core.exceptions import ValidationError
-from core.handlers import (
-    handle_exception,
-    handle_validation_error,
-    save_failed_submission,
-)
-from core.validation.failures import internal, user
-
-
-def test_handle_validation_error_internal(test_session, mocker, caplog):
-    mock_save_failed_submission = mocker.patch("core.handlers.save_failed_submission", return_value="AN ID")
-
-    failures = [internal.ExtraTableFailure(extra_table="Test Table")]
-    with caplog.at_level(logging.ERROR) and test_session.application.test_request_context():
-        file = BytesIO(b"mock file")
-        g.excel_file = file
-        response, status_code = handle_validation_error(ValidationError(validation_failures=failures))
-
-    assert status_code == 500
-    assert response == {
-        "detail": "Internal ingest exception.",
-        "id": "AN ID",
-        "status": 500,
-        "title": "Internal Server Error",
-        "internal_errors": failures,
-    }
-    assert "failure_id" in caplog.text
-    assert "internal_failures" in caplog.text
-    mock_save_failed_submission.assert_called_once()
-
-
-def test_handle_validation_error_user(test_session, mocker):
-    failures = [user.GenericFailure(sheet="Test Sheet", section="Test Section", cell_index="A1", message="Test")]
-    with test_session.application.test_request_context():
-        file = BytesIO(b"mock file")
-        g.excel_file = file
-        response, status_code = handle_validation_error(ValidationError(validation_failures=failures))
-
-    assert status_code == 400
-    assert "detail" in response
-    assert "status" in response
-    assert "title" in response
-    assert "pre_transformation_errors" in response
-    assert "validation_errors" in response
-    assert response["validation_errors"]  # truthy assertion, ignore content
+from core.handlers import handle_exception, save_failed_submission
 
 
 def test_handle_exception_ingest_endpoint(test_session, mocker, caplog):
